Separate test cases to avoid data races on the globals
Tests that re-init the demultiplexer conflict with the tests that do
not, because background goroutines for checks access the demultiplexer
instance when the next test starts.

This isn't an issue during normal agent run, because then the global
demultiplexer instance is read-only and not modified past initialization.
